+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: nappalav <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/23 15:43:38 by nappalav          #+#    #+#             */
-/*   Updated: 2023/11/24 12:40:29 by nappalav         ###   ########.fr       */
+/*   Updated: 2023/11/26 17:18:05 by nappalav         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,6 +24,6 @@ int	ft_printf_c(int c);
 int	ft_printf_s(char *str);
 int	ft_printf_u(unsigned int nbr, char *base);
 int	ft_printf_d(int nbr);
-int	ft_printf_p(unsigned int nbr, char *base);
+int	ft_printf_p(unsigned long nbr, char *base);
 
 #endif
